+++ src/components/Galeria/Galeria.css
@@ -1,23 +1,16 @@
-.galeria_imagen{
+/*.galeria_imagen{
     width:30%;
     height: 30%;
-
+}*/
+/*
+.container-img{
+    flex-direction: column;
+   justify-content: space-between;
 }
-
-.resultContainer{
-    display:flex;
-    justify-content: space-between;
-}
-
-.cardContainer {  
-    margin-top: 1cm;
-    display:flex;
-    justify-content: space-between;
-    }
 
 .detalle_ave{
     background: rgba(12, 11, 11, 0.3);
     color: white;
     font-style: bold;
     height: 30px;
-    }
+    }*/
