+++ src/components/ImagenMediana/ImagenMediana.css
@@ -1,0 +1,4 @@
+/*.imgMediana{
+    max-width: 30%;
+    max-height: 30%;
+}*/
